+++ javascript/startup.js
@@ -2,6 +2,8 @@
 window.api = '/sdapi/v1';
 window.subpath = '';
 
+const startupPromises = [];
+
 async function waitForOpts() {
   // make sure all of the ui is ready and options are loaded
   const t0 = performance.now();
@@ -14,7 +16,8 @@ async function waitForOpts() {
     if (window.opts && Object.keys(window.opts).length > 0) {
       ok = window.opts.theme_type === 'Modern' ? 'uiux_separator_appearance' in window.opts : true;
       if (ok) {
-        log('waitForOpts', `time=${Math.round(t1 - t0)}`);
+        log('waitForOpts', Math.round(t1 - t0));
+        timer('waitForOpts', t1 - t0);
         break;
       }
     }
@@ -23,54 +26,70 @@ async function waitForOpts() {
   }
 }
 
+async function postStartup() {
+  log('postStartup');
+  if (window.gradioObserver) window.gradioObserver.disconnect();
+  if (window.hintsObserver) window.hintsObserver.disconnect();
+  logTimers();
+}
+
 async function initStartup() {
   const t0 = performance.now();
-  log('initGradio', `time=${Math.round(t0 - appStartTime)}`);
+  log('initGradio', Math.round(t0 - appStartTime));
+  timer('initGradio', t0 - appStartTime);
   log('initUi');
   if (window.setupLogger) await setupLogger();
 
   // all items here are non-blocking async calls
-  initModels();
-  getUIDefaults();
-  initPromptChecker();
-  initContextMenu();
-  initDragDrop();
-  initAccordions();
-  initSettings();
-  initImageViewer();
-  initiGenerationParams();
-  initChangelog();
-  setupControlUI();
+
+  startupPromises.push(initModels());
+  startupPromises.push(getUIDefaults());
+  startupPromises.push(initPromptChecker());
+  startupPromises.push(initContextMenu());
+  startupPromises.push(initDragDrop());
+  startupPromises.push(initAccordions());
+  startupPromises.push(initSettings());
+  startupPromises.push(initImageViewer());
+  startupPromises.push(initiGenerationParams());
+  startupPromises.push(initChangelog());
+  startupPromises.push(setupControlUI());
 
   // reconnect server session
   await reconnectUI();
   await waitForOpts();
-  await initGallery();
 
   log('mountURL', window.opts.subpath);
   if (window.opts.subpath?.length > 0) {
     window.subpath = window.opts.subpath;
     window.api = `${window.subpath}/sdapi/v1`;
   }
-  setRefreshInterval();
+
   executeCallbacks(uiReadyCallbacks);
-  setupExtraNetworks();
+  startupPromises.push(initGallery());
+  startupPromises.push(setRefreshInterval());
+  startupPromises.push(setupExtraNetworks());
 
   // optinally wait for modern ui
   if (window.waitForUiReady) await waitForUiReady();
-  initAutocomplete();
-  monitorConnection();
 
   // post startup tasks that may take longer but are not critical
-  showNetworks();
-  setHints();
-  applyStyles();
-  initIndexDB();
-  initLogMonitor();
+  startupPromises.push(initAutocomplete());
+  startupPromises.push(monitorConnection());
+  startupPromises.push(showNetworks());
+  startupPromises.push(setHints());
+  startupPromises.push(applyStyles());
+  startupPromises.push(initIndexDB());
+  startupPromises.push(initLogMonitor());
+
   t1 = performance.now();
   log('initStartup', Math.round(1000 * (t1 - t0) / 1000000));
 
   removeSplash();
+
+  await Promise.all(startupPromises);
+  t2 = performance.now();
+  log('initComplete', Math.round(1000 * (t2 - t0) / 1000000));
+  postStartup();
 }
 
 onUiLoaded(initStartup);